feat(23-01): extract matchFilterValue + add shared filterValueSchema
- Extract pure matchFilterValue from filterEngine (D-08, behavior identical)
- filterEngine imports the shared helper; tests still GREEN
- Add filterValueSchema zod discriminated union (D-02, T-23-01-02 mitigation)
- Bump semver 2.39.0 -> 2.40.0
