+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "worldwideview",
-  "version": "2.39.0",
+  "version": "2.40.0",
   "private": true,
   "type": "module",
   "packageManager": "pnpm@9.15.4",
