+++ src/core/filters/filterEngine.ts
@@ -3,6 +3,7 @@ import type {
     FilterDefinition,
     FilterValue,
 } from "@/core/plugins/PluginTypes";
+import { matchFilterValue } from "./matchFilterValue";
 
 /**
  * Apply active filters to a list of entities.
@@ -27,35 +28,10 @@ export function applyFilters(
 
             const propValue = entity.properties[def.propertyKey];
 
-            if (!matchesFilter(propValue, filterVal)) {
+            if (!matchFilterValue(propValue, filterVal)) {
                 return false;
             }
         }
         return true;
     });
-}
-
-function matchesFilter(propValue: unknown, filter: FilterValue): boolean {
-    switch (filter.type) {
-        case "text": {
-            if (!filter.value) return true; // empty text = no filter
-            const str = String(propValue ?? "").toLowerCase();
-            return str.includes(filter.value.toLowerCase());
-        }
-        case "select": {
-            if (filter.values.length === 0) return true; // no selection = no filter
-            const pVal = String(propValue ?? "").toLowerCase();
-            return filter.values.some((v) => v.toLowerCase() === pVal);
-        }
-        case "range": {
-            const num = Number(propValue ?? 0);
-            if (isNaN(num)) return false;
-            return num >= filter.min && num <= filter.max;
-        }
-        case "boolean": {
-            return Boolean(propValue) === filter.value;
-        }
-        default:
-            return true;
-    }
 }
